+++ src/bullet.hpp
@@ -11,7 +11,7 @@ public:
 	Bullet(float width, float height);
 	void drawTo(sf::RenderWindow& window);
 	void setPos(float player_x, float player_y);
-	void fireBullet(Player& player_rect, float bullet_speed);
+	void fireBullet(Player& player_rect, int bullet_speed);
 	void bulletCollision(Dinosaurs& dino_rect, bool& dino_dead, bool& bullet_dead, int& score, float screen_width, float screen_height);
 	void moveBullet();
 	float returnX();
